Validate Employee Abbreviation before save
