+++ coverage.out
@@ -1,4 +1,5 @@
 mode: set
+greenbone-case-study/docs/docs.go:451.13,453.2 1 0
 greenbone-case-study/main.go:16.13,27.2 6 0
 greenbone-case-study/database/db.go:14.18,19.16 2 0
 greenbone-case-study/database/db.go:19.16,21.3 1 0
@@ -6,29 +7,30 @@ greenbone-case-study/database/db.go:23.2,24.15 2 0
 greenbone-case-study/database/db.go:27.35,32.16 2 0
 greenbone-case-study/database/db.go:32.16,34.3 1 0
 greenbone-case-study/database/db.go:35.2,36.17 2 0
-greenbone-case-study/docs/docs.go:401.13,403.2 1 0
 greenbone-case-study/models/computer.go:25.36,26.69 1 0
 greenbone-case-study/models/computer.go:26.69,28.3 1 0
 greenbone-case-study/models/computer.go:29.2,29.13 1 0
 greenbone-case-study/models/computer.go:32.56,33.19 1 0
 greenbone-case-study/models/computer.go:33.19,35.3 1 0
-greenbone-case-study/models/computer.go:36.2,40.15 3 0
-greenbone-case-study/models/computer.go:40.15,42.3 1 0
-greenbone-case-study/models/computer.go:44.2,45.15 2 0
-greenbone-case-study/models/computer.go:45.15,47.3 1 0
-greenbone-case-study/models/computer.go:49.2,49.12 1 0
-greenbone-case-study/models/computer.go:52.55,54.20 2 0
-greenbone-case-study/models/computer.go:54.20,56.3 1 0
-greenbone-case-study/models/computer.go:57.2,60.16 3 0
-greenbone-case-study/models/computer.go:60.16,62.3 1 0
-greenbone-case-study/models/computer.go:64.2,64.12 1 0
-greenbone-case-study/models/computer.go:70.53,85.16 4 0
-greenbone-case-study/models/computer.go:85.16,87.3 1 0
-greenbone-case-study/models/computer.go:89.2,91.16 2 0
-greenbone-case-study/models/computer.go:91.16,93.3 1 0
-greenbone-case-study/models/computer.go:95.2,97.38 2 0
-greenbone-case-study/models/computer.go:97.38,99.3 1 0
-greenbone-case-study/models/computer.go:100.2,100.12 1 0
+greenbone-case-study/models/computer.go:36.2,39.37 2 0
+greenbone-case-study/models/computer.go:39.37,41.3 1 0
+greenbone-case-study/models/computer.go:43.2,44.15 2 0
+greenbone-case-study/models/computer.go:44.15,46.3 1 0
+greenbone-case-study/models/computer.go:48.2,49.15 2 0
+greenbone-case-study/models/computer.go:49.15,51.3 1 0
+greenbone-case-study/models/computer.go:53.2,53.12 1 0
+greenbone-case-study/models/computer.go:56.55,58.20 2 0
+greenbone-case-study/models/computer.go:58.20,60.3 1 0
+greenbone-case-study/models/computer.go:61.2,64.16 3 0
+greenbone-case-study/models/computer.go:64.16,66.3 1 0
+greenbone-case-study/models/computer.go:68.2,68.12 1 0
+greenbone-case-study/models/computer.go:74.53,89.16 4 0
+greenbone-case-study/models/computer.go:89.16,91.3 1 0
+greenbone-case-study/models/computer.go:93.2,95.16 2 0
+greenbone-case-study/models/computer.go:95.16,97.3 1 0
+greenbone-case-study/models/computer.go:99.2,101.38 2 0
+greenbone-case-study/models/computer.go:101.38,103.3 1 0
+greenbone-case-study/models/computer.go:104.2,104.12 1 0
 greenbone-case-study/routes/route.go:12.45,13.40 1 0
 greenbone-case-study/routes/route.go:13.40,15.3 1 0
 greenbone-case-study/routes/route.go:17.2,17.46 1 0
@@ -89,20 +91,22 @@ greenbone-case-study/models/computer.go:26.69,28.3 1 1
 greenbone-case-study/models/computer.go:29.2,29.13 1 1
 greenbone-case-study/models/computer.go:32.56,33.19 1 1
 greenbone-case-study/models/computer.go:33.19,35.3 1 1
-greenbone-case-study/models/computer.go:36.2,40.15 3 1
-greenbone-case-study/models/computer.go:40.15,42.3 1 1
-greenbone-case-study/models/computer.go:44.2,45.15 2 1
-greenbone-case-study/models/computer.go:45.15,47.3 1 1
-greenbone-case-study/models/computer.go:49.2,49.12 1 1
-greenbone-case-study/models/computer.go:52.55,54.20 2 1
-greenbone-case-study/models/computer.go:54.20,56.3 1 1
-greenbone-case-study/models/computer.go:57.2,60.16 3 1
-greenbone-case-study/models/computer.go:60.16,62.3 1 1
-greenbone-case-study/models/computer.go:64.2,64.12 1 1
-greenbone-case-study/models/computer.go:70.53,85.16 4 1
-greenbone-case-study/models/computer.go:85.16,87.3 1 0
-greenbone-case-study/models/computer.go:89.2,91.16 2 1
-greenbone-case-study/models/computer.go:91.16,93.3 1 0
-greenbone-case-study/models/computer.go:95.2,97.38 2 1
-greenbone-case-study/models/computer.go:97.38,99.3 1 0
-greenbone-case-study/models/computer.go:100.2,100.12 1 1
+greenbone-case-study/models/computer.go:36.2,39.37 2 1
+greenbone-case-study/models/computer.go:39.37,41.3 1 0
+greenbone-case-study/models/computer.go:43.2,44.15 2 1
+greenbone-case-study/models/computer.go:44.15,46.3 1 1
+greenbone-case-study/models/computer.go:48.2,49.15 2 1
+greenbone-case-study/models/computer.go:49.15,51.3 1 1
+greenbone-case-study/models/computer.go:53.2,53.12 1 1
+greenbone-case-study/models/computer.go:56.55,58.20 2 1
+greenbone-case-study/models/computer.go:58.20,60.3 1 1
+greenbone-case-study/models/computer.go:61.2,64.16 3 1
+greenbone-case-study/models/computer.go:64.16,66.3 1 1
+greenbone-case-study/models/computer.go:68.2,68.12 1 1
+greenbone-case-study/models/computer.go:74.53,89.16 4 1
+greenbone-case-study/models/computer.go:89.16,91.3 1 0
+greenbone-case-study/models/computer.go:93.2,95.16 2 1
+greenbone-case-study/models/computer.go:95.16,97.3 1 0
+greenbone-case-study/models/computer.go:99.2,101.38 2 1
+greenbone-case-study/models/computer.go:101.38,103.3 1 0
+greenbone-case-study/models/computer.go:104.2,104.12 1 1
