+++ handlers/computers/assign_test.go
@@ -13,8 +13,8 @@ import (
 
 func TestAssignComputer(t *testing.T) {
 	app := setupTestApp()
-	computer := createComputer("TestPC", "11:22:33:44:55:90", "192.168.1.2", "EMP1")
-	req := httptest.NewRequest("PUT", fmt.Sprintf("/api/computers/%d/assign/EMP2", computer.ID), nil)
+	computer := createComputer("TestPC", "11:22:33:44:55:90", "192.168.1.2", "EM1")
+	req := httptest.NewRequest("PUT", fmt.Sprintf("/api/computers/%d/assign/EM2", computer.ID), nil)
 	resp, err := app.Test(req)
 	assert.Nil(t, err)
 	assert.Equal(t, 200, resp.StatusCode)
@@ -22,12 +22,12 @@ func TestAssignComputer(t *testing.T) {
 	body, _ := io.ReadAll(resp.Body)
 	err = json.Unmarshal(body, &updatedComputer)
 	assert.Nil(t, err)
-	assert.Equal(t, "EMP2", updatedComputer.EmployeeAbbreviation)
+	assert.Equal(t, "EM2", updatedComputer.EmployeeAbbreviation)
 }
 func TestAssignComputerNotFound(t *testing.T) {
 	app := setupTestApp()
 	fakeID := 9999
-	req := httptest.NewRequest("PUT", fmt.Sprintf("/api/computers/%d/assign/EMP2", fakeID), nil)
+	req := httptest.NewRequest("PUT", fmt.Sprintf("/api/computers/%d/assign/EM3", fakeID), nil)
 	resp, err := app.Test(req)
 	assert.Nil(t, err)
 	assert.Equal(t, 404, resp.StatusCode)
